+++ R/expected_dist.R
@@ -11,7 +11,6 @@
 #' @example /inst/examples/expected_dist_example.R
 
 expected_dist <- function(alpha, n_items, metric) {
-
   if (n_items < 1 | floor(n_items) != n_items) {
     stop("Number of items must be a positive integer")
   }
@@ -21,7 +20,7 @@ expected_dist <- function(alpha, n_items, metric) {
 
   if (alpha < 0) {
     stop("alpha must be a non-negative value")
-  }else{
+  } else {
     if (metric == "kendall") {
       out <- exp_d_tau(alpha, n_items)
     }
@@ -32,12 +31,15 @@ expected_dist <- function(alpha, n_items, metric) {
       out <- exp_d_ham(alpha, n_items)
     }
     if (metric %in% c("ulam", "footrule", "spearman")) {
-      pfd <- dplyr::filter(partition_function_data,
-                           .data$metric == !!metric, .data$n_items == !!n_items,
-                           .data$type == "cardinalities")
+      pfd <- partition_function_data[
+        partition_function_data$metric == metric &
+          partition_function_data$n_items == n_items &
+          partition_function_data$type == "cardinalities", ,
+        drop = FALSE
+      ]
       if (nrow(pfd) == 0) {
         stop("Given number of items currently not available for the specified metric")
-      } else{
+      } else {
         card <- pfd$values[[1]]
       }
 
@@ -47,7 +49,9 @@ expected_dist <- function(alpha, n_items, metric) {
           alpha = alpha * n_items,
           n_items = n_items,
           cardinalities = card,
-          metric = metric))
+          metric = metric
+        )
+      )
     }
   }
   return(out)