+++ R/tidy_mcmc.R
@@ -1,5 +1,4 @@
 tidy_mcmc <- function(fit) {
-
   fit <- tidy_rho(fit)
   fit <- tidy_alpha(fit)
   fit <- tidy_cluster_assignment(fit)
@@ -31,13 +30,15 @@ tidy_rho <- function(fit) {
     times = rho_dims[[3]]
   )
   cluster <- factor(paste("Cluster", cluster),
-                    levels = paste("Cluster", sort(unique(cluster))))
+    levels = paste("Cluster", sort(unique(cluster)))
+  )
 
   iteration <- rep(seq(from = 1, to = rho_dims[[3]] * fit$rho_thinning, by = fit$rho_thinning),
-                   each = rho_dims[[1]] * rho_dims[[2]])
+    each = rho_dims[[1]] * rho_dims[[2]]
+  )
 
-  # Store the final rho as a tibble
-  fit$rho <- dplyr::tibble(
+  # Store the final rho as a dataframe
+  fit$rho <- data.frame(
     item = item,
     cluster = cluster,
     iteration = iteration,
@@ -60,14 +61,15 @@ tidy_alpha <- function(fit) {
   )
 
   cluster <- factor(paste("Cluster", cluster),
-                    levels = paste("Cluster", sort(unique(cluster))))
+    levels = paste("Cluster", sort(unique(cluster)))
+  )
 
   iteration <- rep(
     seq(from = 1, to = alpha_dims[[2]] * fit$alpha_jump, by = fit$alpha_jump),
     each = alpha_dims[[1]]
   )
 
-  fit$alpha <- dplyr::tibble(
+  fit$alpha <- data.frame(
     cluster = cluster,
     iteration = iteration,
     value = value
@@ -101,7 +103,7 @@ tidy_cluster_assignment <- function(fit) {
       each = cluster_dims[[1]]
     )
 
-    fit$cluster_assignment <- dplyr::tibble(
+    fit$cluster_assignment <- data.frame(
       assessor = assessor,
       iteration = iteration,
       value = value
@@ -133,14 +135,15 @@ tidy_cluster_probabilities <- function(fit) {
   )
 
   cluster <- factor(paste("Cluster", cluster),
-                    levels = paste("Cluster", sort(unique(cluster))))
+    levels = paste("Cluster", sort(unique(cluster)))
+  )
 
   iteration <- rep(
     seq(from = 1, to = clusprob_dims[[2]], by = 1),
     each = clusprob_dims[[1]]
   )
 
-  fit$cluster_probs <- dplyr::tibble(
+  fit$cluster_probs <- data.frame(
     cluster = cluster,
     iteration = iteration,
     value = value
@@ -163,20 +166,19 @@ tidy_wcd <- function(fit) {
       times = wcd_dims[[2]]
     )
     cluster <- factor(paste("Cluster", cluster),
-                      levels = paste("Cluster", sort(unique(cluster))))
+      levels = paste("Cluster", sort(unique(cluster)))
+    )
 
     iteration <- rep(
       seq(from = 1, to = wcd_dims[[2]], by = 1),
       each = wcd_dims[[1]]
     )
 
-    fit$within_cluster_distance <- dplyr::tibble(
+    fit$within_cluster_distance <- data.frame(
       cluster = cluster,
       iteration = iteration,
       value = value
     )
-
-
   } else {
     fit$within_cluster_distance <- NULL
   }
@@ -186,7 +188,6 @@ tidy_wcd <- function(fit) {
 tidy_augmented_data <- function(fit) {
   # Tidy augmented data, or delete
   if (fit$save_aug) {
-
     augdata_dims <- dim(fit$augmented_data)
 
     # Item1, Item2, ..., Item1, Item2, ..., Item1, Item2, ..., Item1, Item2
@@ -197,13 +198,16 @@ tidy_augmented_data <- function(fit) {
     item <- rep(fit$items, times = augdata_dims[[2]] * augdata_dims[[3]])
     item <- factor(item, levels = fit$items)
 
-    assessor <- rep(seq(from = 1, to = augdata_dims[[2]], by = 1), each = augdata_dims[[1]],
-                    times = augdata_dims[[3]])
+    assessor <- rep(seq(from = 1, to = augdata_dims[[2]], by = 1),
+      each = augdata_dims[[1]],
+      times = augdata_dims[[3]]
+    )
 
     iteration <- rep(seq(from = 1, to = augdata_dims[[3]] * fit$aug_thinning, by = fit$aug_thinning),
-                     each = augdata_dims[[1]] * augdata_dims[[2]])
+      each = augdata_dims[[1]] * augdata_dims[[2]]
+    )
 
-    fit$augmented_data <- dplyr::tibble(
+    fit$augmented_data <- data.frame(
       iteration = iteration,
       assessor = assessor,
       item = item,
@@ -221,12 +225,9 @@ tidy_augmentation_acceptance <- function(fit) {
   # Augmentation acceptance
 
   if (fit$any_missing || fit$augpair) {
-    fit$aug_acceptance <- dplyr::tibble(acceptance_rate = c(fit$aug_acceptance))
-    fit$aug_acceptance <- dplyr::mutate(fit$aug_acceptance,
-                                        assessor = dplyr::row_number())
-    fit$aug_acceptance <- dplyr::select(fit$aug_acceptance,
-                                        .data$assessor, .data$acceptance_rate)
-
+    fit$aug_acceptance <- data.frame(acceptance_rate = c(fit$aug_acceptance))
+    fit$aug_acceptance$assessor <- seq_len(nrow(fit$aug_acceptance))
+    fit$aug_acceptance <- fit$aug_acceptance[, c("assessor", "acceptance_rate")]
   } else {
     fit$aug_acceptance <- NULL
   }
@@ -239,7 +240,7 @@ tidy_error_probability <- function(fit) {
   theta_length <- length(fit$theta)
 
   if (theta_length > 0) {
-    fit$theta <- dplyr::tibble(
+    fit$theta <- data.frame(
       iteration = seq(from = 1, to = theta_length, by = 1),
       value = c(fit$theta)
     )