+++ inst/examples/label_switching_example.R
@@ -47,16 +47,13 @@
       read.csv(cluster_files[[i]], header = FALSE))
   }
 
-  library(dplyr)
+
   # Create an integer array of latent allocations, as this is required by label.switching
-  z <- m$cluster_assignment %>%
-    filter(iteration > burnin) %>%
-    mutate(value = as.integer(gsub("Cluster ", "", value))) %>%
-    as.data.frame() %>%
-    stats::reshape(direction = "wide",
-                   idvar = "iteration", timevar = "assessor") %>%
-    select(-iteration) %>%
-    as.matrix()
+  z <- subset(m$cluster_assignment, iteration > burnin)
+  z$value <- as.integer(gsub("Cluster ", "", z$value))
+  z <- stats::reshape(z, direction = "wide", idvar = "iteration", timevar = "assessor")
+  z$iteration <- NULL
+  z <- as.matrix(z)
 
   # Now apply Stephen's algorithm
   library(label.switching)