+++ tests/testthat/test-partition_function.R
@@ -1,6 +1,5 @@
 context("Testing computation of partition functions")
 
-
 # Brute force formula
 check_log_zn <- function(n, alpha, metric) {
   # Generate all permutations
@@ -9,21 +8,24 @@ check_log_zn <- function(n, alpha, metric) {
 
   # Compute the partition function
   if (metric == "footrule") {
-    log(sum(exp(- alpha / n * colSums(abs(t(perm) - 1:n)))))
+    log(sum(exp(-alpha / n * colSums(abs(t(perm) - 1:n)))))
   } else if (metric == "spearman") {
-    log(sum(exp(- alpha / n * colSums((t(perm) - 1:n)^2))))
+    log(sum(exp(-alpha / n * colSums((t(perm) - 1:n)^2))))
   } else if (metric == "kendall") {
-    log(sum(exp(- alpha / n * apply(perm, 1,
-                                    get_rank_distance,
-                                    r2 = 1:n, metric = "kendall"))))
+    log(sum(exp(-alpha / n * apply(perm, 1,
+      get_rank_distance,
+      r2 = 1:n, metric = "kendall"
+    ))))
   } else if (metric == "cayley") {
-    log(sum(exp(- alpha / n * apply(perm, 1,
-                                    get_rank_distance,
-                                    r2 = 1:n, metric = "cayley"))))
+    log(sum(exp(-alpha / n * apply(perm, 1,
+      get_rank_distance,
+      r2 = 1:n, metric = "cayley"
+    ))))
   } else if (metric == "hamming") {
-    log(sum(exp(- alpha / n * apply(perm, 1,
-                                    get_rank_distance,
-                                    r2 = 1:n, metric = "hamming"))))
+    log(sum(exp(-alpha / n * apply(perm, 1,
+      get_rank_distance,
+      r2 = 1:n, metric = "hamming"
+    ))))
   } else if (metric == "ulam") {
     log(sum(unlist(lapply(seq(0, n - 1, by = 1), function(x) {
       PerMallows::count.perms(perm.length = n, dist.value = x, dist.name = "ulam") * exp(-alpha / n * x)
@@ -35,34 +37,42 @@ check_log_zn <- function(n, alpha, metric) {
 
 # Loop over some n and alpha values
 test_that("footrule partition function is correct", {
-
-  footrule_sequence <- dplyr::filter(BayesMallows:::partition_function_data,
-                                     metric == "footrule", type == "cardinalities")$values
+  footrule_sequence <- subset(
+    BayesMallows:::partition_function_data,
+    metric == "footrule" & type == "cardinalities"
+  )$values
 
   for (n in c(1, 2, 3, 5)) {
     for (alpha in c(0.001, 0.1, 1)) {
       expect_equal(
-        get_partition_function(n = n, alpha = alpha,
-                               cardinalities = footrule_sequence[[n]], metric = "footrule"),
+        get_partition_function(
+          n = n, alpha = alpha,
+          cardinalities = footrule_sequence[[n]], metric = "footrule"
+        ),
         check_log_zn(n, alpha, "footrule")
       )
     }
-  }})
+  }
+})
 
 test_that("Spearman partition function is correct", {
-
-  spearman_sequence <- dplyr::filter(BayesMallows:::partition_function_data,
-                  metric == "spearman", type == "cardinalities")$values
+  spearman_sequence <- subset(
+    BayesMallows:::partition_function_data,
+    metric == "spearman" & type == "cardinalities"
+  )$values
 
   for (n in c(1, 2, 3)) {
     for (alpha in c(0.001, 0.1, 1)) {
       expect_equal(
-        get_partition_function(n = n, alpha = alpha,
-                               cardinalities = spearman_sequence[[n]], metric = "spearman"),
+        get_partition_function(
+          n = n, alpha = alpha,
+          cardinalities = spearman_sequence[[n]], metric = "spearman"
+        ),
         check_log_zn(n, alpha, "spearman")
       )
     }
-  }})
+  }
+})
 
 
 test_that("Kendall partition function is correct", {
@@ -73,7 +83,8 @@ test_that("Kendall partition function is correct", {
         check_log_zn(n, alpha, "kendall")
       )
     }
-  }})
+  }
+})
 
 test_that("Cayley partition function is correct", {
   for (n in c(1, 2, 3)) {
@@ -83,7 +94,8 @@ test_that("Cayley partition function is correct", {
         check_log_zn(n, alpha, "cayley")
       )
     }
-  }})
+  }
+})
 
 
 test_that("Hamming partition function is correct", {
@@ -94,32 +106,39 @@ test_that("Hamming partition function is correct", {
         check_log_zn(n, alpha, "hamming")
       )
     }
-  }})
+  }
+})
 
 test_that("Ulam partition function is correct", {
-
-  ulam_sequence <- dplyr::filter(BayesMallows:::partition_function_data,
-                                 metric == "ulam", type == "cardinalities")$values
+  ulam_sequence <- subset(
+    BayesMallows:::partition_function_data,
+    metric == "ulam" & type == "cardinalities"
+  )$values
   for (n in c(1, 2, 3)) {
     for (alpha in c(0.001, 0.1, 1)) {
       expect_equal(
-        get_partition_function(n = n, alpha = alpha,
-                               cardinalities = ulam_sequence[[n]],
-                               metric = "ulam"),
+        get_partition_function(
+          n = n, alpha = alpha,
+          cardinalities = ulam_sequence[[n]],
+          metric = "ulam"
+        ),
         check_log_zn(n, alpha, "ulam")
       )
     }
-  }})
+  }
+})
 
 
 test_that("partition function data is sane", {
+  pfd <- aggregate(
+    list(n = seq_len(nrow(BayesMallows:::partition_function_data))),
+    BayesMallows:::partition_function_data[, c("n_items", "metric", "type")],
+    FUN = length)
+
   expect_equal(
-    BayesMallows:::partition_function_data %>%
-      group_by(n_items, metric, type) %>%
-      count() %>%
-      filter(n > 1) %>%
-      nrow(),
-    0)
+    nrow(subset(pfd, n > 1)),
+    0
+  )
 })
 
 test_that("estimate_partition_function runs in parallel", {
@@ -128,22 +147,23 @@ test_that("estimate_partition_function runs in parallel", {
   metric <- "spearman"
   degree <- 2
 
-  fit <- estimate_partition_function(method = "importance_sampling",
-                                     alpha_vector = alpha_vector,
-                                     n_items = n_items, metric = metric,
-                                     nmc = 20, degree = degree)
+  fit <- estimate_partition_function(
+    method = "importance_sampling",
+    alpha_vector = alpha_vector,
+    n_items = n_items, metric = metric,
+    nmc = 20, degree = degree
+  )
 
 
   lapply(1:2, function(x) {
     cl <- parallel::makeCluster(x)
-    fit <- estimate_partition_function(method = "importance_sampling",
-                                       alpha_vector = alpha_vector,
-                                       n_items = n_items, metric = metric,
-                                       nmc = 20, degree = degree, cl = cl)
+    fit <- estimate_partition_function(
+      method = "importance_sampling",
+      alpha_vector = alpha_vector,
+      n_items = n_items, metric = metric,
+      nmc = 20, degree = degree, cl = cl
+    )
     parallel::stopCluster(cl)
     expect_equal(length(fit), 3)
   })
-
-
-
 })