+++ tests/testthat/test-smc_mallows_partial_rankings.R
@@ -3,16 +3,16 @@ context("SMC new users partial rankings")
 # Generate Dataset =======================================
 
 # General ------------------------------------------------
-n_items   <- dim(sushi_rankings)[2] # Number of items
-rho_0     <- seq(from = 1, to = n_items, by = 1) # 'true' consensus ranking
-alpha_0   <- 2 # fixed/ 'true' scale parameter
+n_items <- dim(sushi_rankings)[2] # Number of items
+rho_0 <- seq(from = 1, to = n_items, by = 1) # 'true' consensus ranking
+alpha_0 <- 2 # fixed/ 'true' scale parameter
 leap_size <- floor(n_items / 5)
-metric    <- "footrule"
+metric <- "footrule"
 
 # Generate estimate of Z_n(alpha) ------------------------
 alpha_vector <- seq(from = 0, to = 20, by = 1)
-iter         <- 1e2
-degree       <- 10
+iter <- 1e2
+degree <- 10
 
 # Estimate the logarithm of the partition function of the Mallows rank model
 # using the estimate partition function
@@ -27,7 +27,7 @@ logz_estimate <- estimate_partition_function(
 )
 
 # Make this information partially observed over time -----
-samples              <- sushi_rankings[1: 100, ]
+samples <- sushi_rankings[1:100, ]
 samples[samples > 5] <- NA
 
 # Bayesmallows MCMC Results ==============================
@@ -56,7 +56,7 @@ test_that("BayesMallows MCMC Results are OK", {
     type      = "MAP",
     burnin    = bm_mcmc$burnin
   )
-  post_rho   <- compute_posterior_intervals(bm_mcmc, parameter = "rho")
+  post_rho <- compute_posterior_intervals(bm_mcmc, parameter = "rho")
   post_alpha <- compute_posterior_intervals(bm_mcmc, parameter = "alpha")
   expect_equal(dim(post_rho), c(10, 7))
   expect_equal(dim(rho_cp), c(10, 3))
@@ -66,12 +66,12 @@ test_that("BayesMallows MCMC Results are OK", {
 
 # SMC Analysis (alpha unknown) ===========================
 
-mcmc_times    <- 5
-num_new_obs   <- 5
-Time          <- dim(samples)[1] / num_new_obs
+mcmc_times <- 5
+num_new_obs <- 5
+Time <- dim(samples)[1] / num_new_obs
 alpha_prop_sd <- 0.5
-lambda        <- 0.15
-alpha_max     <- 1e0
+lambda <- 0.15
+alpha_max <- 1e0
 
 test_that("Produces the wrong metric and aug_method error", {
   N <- 5
@@ -124,7 +124,7 @@ test_that("Runs with unif kernel", {
     aug_method      = "random"
   )
   expect_is(smc_unif_alpha_fixed_unif, "list")
-  expect_equal(length(smc_unif_alpha_fixed_unif), 1)
+  expect_equal(length(smc_unif_alpha_fixed_unif), 3)
   expect_equal(dim(smc_unif_alpha_fixed_unif$rho_samples), c(N, 10, 21))
   smc_unif <- smc_mallows_new_users_partial(
     R_obs           = samples,
@@ -142,17 +142,19 @@ test_that("Runs with unif kernel", {
     aug_method      = "random"
   )
   expect_is(smc_unif, "list")
-  expect_equal(length(smc_unif), 2)
+  expect_equal(length(smc_unif), 4)
   expect_equal(dim(smc_unif$rho_samples), c(N, 10, 21))
   expect_equal(dim(smc_unif$alpha_samples), c(N, 21))
 
   expect_s3_class(
     plot_alpha_posterior(smc_unif$alpha_samples[, Time + 1], nmc = N, burnin = 2),
-    "ggplot")
+    "ggplot"
+  )
 
   expect_s3_class(
     plot_rho_posterior(smc_unif$rho_samples[, , Time + 1], nmc = N, burnin = 2, C = 1),
-    "ggplot")
+    "ggplot"
+  )
 })
 
 test_that("Runs with pseudo kernel", {
@@ -171,7 +173,7 @@ test_that("Runs with pseudo kernel", {
     aug_method      = "pseudolikelihood"
   )
   expect_is(smc_unif_alpha_fixed_pseudo, "list")
-  expect_equal(length(smc_unif_alpha_fixed_pseudo), 1)
+  expect_equal(length(smc_unif_alpha_fixed_pseudo), 3)
   expect_equal(dim(smc_unif_alpha_fixed_pseudo$rho_samples), c(N, 10, 21))
   smc_pseudo <- smc_mallows_new_users_partial(
     R_obs           = samples,
@@ -189,7 +191,7 @@ test_that("Runs with pseudo kernel", {
     aug_method      = "pseudolikelihood"
   )
   expect_is(smc_pseudo, "list")
-  expect_equal(length(smc_pseudo), 2)
+  expect_equal(length(smc_pseudo), 4)
   expect_equal(dim(smc_pseudo$rho_samples), c(N, 10, 21))
   expect_equal(dim(smc_pseudo$alpha_samples), c(N, 21))
 })
@@ -216,10 +218,27 @@ test_that("Specific example results are OK", {
     alpha_max       = alpha_max,
     aug_method      = aug_method
   )
+
+  expect_error(
+    compute_rho_consensus(output = test$rho_samples[, , Time + 1], nmc = N, burnin = NULL, C = 1,
+                          type = "CP"),
+    "Please specify the burnin."
+    )
+
+  expect_error(
+    compute_rho_consensus(
+      output = test$rho_samples[, , Time + 1], nmc = N, burnin = NULL, C = 1,
+      type = "MAP"
+    ),
+    "Please specify the burnin."
+  )
+
+
   rho_cp <- compute_rho_consensus(
     output = test$rho_samples[, , Time + 1], nmc = N, burnin = 0, C = 1,
     type = "CP"
   )
+
   set.seed(545)
   rho_map <- compute_rho_consensus(
     output = test$rho_samples[, , Time + 1], nmc = N, burnin = 0, C = 1,
@@ -235,6 +254,24 @@ test_that("Specific example results are OK", {
   expect_equal(dim(rho_map), c(30, 3))
   expect_equal(dim(post_rho), c(10, 7))
   expect_equal(dim(post_alpha), c(1, 6))
+
+  # Test with nonzero burnin
+  rho_cp <- compute_rho_consensus(
+    output = test$rho_samples[, , Time + 1], nmc = N, burnin = 2, C = 1,
+    type = "CP"
+  )
+  expect_equal(rho_cp$cumprob,
+               c(1, 0.75, 0.5, 0.5, 0.125, 0.125, 0.375, 0.625, 0.5, 0.5))
+
+  rho_map <- compute_rho_consensus(
+    output = test$rho_samples[, , Time + 1], nmc = N, burnin = 2, C = 1,
+    type = "MAP"
+  )
+  expect_equal(
+    rho_map$probability,
+    c(0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25,
+      0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25)
+  )
 
   test_fixed <- smc_mallows_new_users_partial_alpha_fixed(
     R_obs           = samples,
@@ -264,4 +301,5 @@ test_that("Specific example results are OK", {
   expect_equal(dim(rho_cp_fixed), c(10, 3))
   expect_equal(dim(rho_map_fixed), c(10, 3))
   expect_equal(dim(post_rho_fixed), c(10, 7))
+
 })